debt - more layering fixes
